feat: removed nodejs backend-only items
