+++ README.md
@@ -1,4 +1,4 @@
-# @grayarrow/core-utils
+# @digicrew/core-utils
 
 Provides basic helper utilities for supporting TypeScript and JavaScript code.
