+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@grayarrow/core-utils",
-  "version": "2.1.0",
+  "name": "@digicrew/core-utils",
+  "version": "2.2.0",
   "description": "Core helper utilities for supporting TypeScript and JavaScript projects.",
   "private": false,
   "module": "src/index.mjs",
@@ -14,7 +14,7 @@
     "build:test:base": "rimraf dist && tsc -p tsconfig.test.json",
     "clean": "npm run clean:files && npm install",
     "clean:files": "rimraf .next _next build dist lib node_modules out package-lock.json pnpm-lock.yaml yarn.lock",
-    "export:package": "echo '{\"name\": \"@grayarrow/core-utils\", \"type\": \"module\", \"exports\": \"./dist/index.mjs\", \"version\": \"2.1.0\"}' > dist/package.json",
+    "export:package": "echo '{\"name\": \"@digicrew/core-utils\", \"type\": \"module\", \"exports\": \"./dist/index.mjs\", \"version\": \"2.1.0\"}' > dist/package.json",
     "lint": "eslint ./",
     "lint:fix": "eslint ./ --fix",
     "test": "node --experimental-vm-modules node_modules/.bin/jest",
@@ -42,10 +42,7 @@
   },
   "homepage": "https://github.com/David-Dietrich-wins/core-utils#readme",
   "dependencies": {
-    "jsonwebtoken": "^9.0.2",
     "moment": "^2.30.1",
-    "winston": "^3.17.0",
-    "winston-daily-rotate-file": "^5.0.0",
     "zod": "*"
   },
   "devDependencies": {
