+++ src/jest.setup.mts
@@ -1,9 +1,6 @@
 import * as z from 'zod/v4'
 import { jest } from '@jest/globals'
 // eslint-disable-next-line sort-imports
-import { JwtPayload } from 'jsonwebtoken'
-import { JwtTokenWithEmail } from './services/jwt.mjs'
-import { LogManagerOptions } from './services/LogManager.mjs'
 import { NumberToString } from './primitives/number-helper.mjs'
 import { safestr } from './primitives/string-helper.mjs'
 
@@ -151,36 +148,36 @@ export class ZodTestHelper {
   }
 }
 
-const globalLogger = jest.fn().mockImplementation(() => ({
-  debug: mockLoggerDebug,
-  error: mockLoggerError,
-  info: mockLoggerInfo,
-  log: mockLoggerLog,
-  silly: mockLoggerSilly,
-  warn: mockLoggerWarn,
-}))
-jest.unstable_mockModule('./services/LogManager.mjs', () => ({
-  LogManager: globalLogger,
-}))
+// const globalLogger = jest.fn().mockImplementation(() => ({
+//   debug: mockLoggerDebug,
+//   error: mockLoggerError,
+//   info: mockLoggerInfo,
+//   log: mockLoggerLog,
+//   silly: mockLoggerSilly,
+//   warn: mockLoggerWarn,
+// }))
+// jest.unstable_mockModule('./services/LogManager.mjs', () => ({
+//   LogManager: globalLogger,
+// }))
 
-const { LogManager } = await import('./services/LogManager.mjs')
-export function getGlobalLogger() {
-  const loggerOptions: LogManagerOptions = {
-    componentName: 'test',
-    includeHttpRequestDataInTheLog: true,
-    includeHttpResponseDataInTheLog: true,
-    logBaseFileName: 'test',
-    logFileName: 'test.log',
-    logLevel: 'all',
-    maxFiles: 10,
-    maxSize: 1000000,
-    rotateBaseFileName: 'test',
-    showConsole: true,
-    suffixDatePattern: 'YYYY-MM-DD-HH',
-  }
+// const { LogManager } = await import('./services/LogManager.mjs')
+// export function getGlobalLogger() {
+//   const loggerOptions: LogManagerOptions = {
+//     componentName: 'test',
+//     includeHttpRequestDataInTheLog: true,
+//     includeHttpResponseDataInTheLog: true,
+//     logBaseFileName: 'test',
+//     logFileName: 'test.log',
+//     logLevel: 'all',
+//     maxFiles: 10,
+//     maxSize: 1000000,
+//     rotateBaseFileName: 'test',
+//     showConsole: true,
+//     suffixDatePattern: 'YYYY-MM-DD-HH',
+//   }
 
-  return new LogManager(loggerOptions)
-}
+//   return new LogManager(loggerOptions)
+// }
 
 // Const httpHandlers: HttpHandler[] = []
 
@@ -192,7 +189,7 @@ export const TEST_Settings: {
   currentDate: Date
   currentDateInMilliseconds: number
   currentDateString: string
-  jwt: string
+  // jwt: string
   rsaPassPhrase: string
   rsaPrivateKey: string
   rsaPublicKey: string
@@ -211,12 +208,12 @@ export const TEST_Settings: {
     TEST_Settings.rsaPublicKey = safestr(process.env.rsaPublicKey)
 
     // eslint-disable-next-line no-use-before-define
-    TEST_Settings.jwt = GenerateSignedJwtToken(TEST_Settings.userIdGoodEmail)
+    // TEST_Settings.jwt = GenerateSignedJwtToken(TEST_Settings.userIdGoodEmail)
   },
   currentDate: new Date('2025-12-01T12:00:00.000Z'),
   currentDateInMilliseconds: 0,
   currentDateString: '2025-12-01T12:00:00.000Z',
-  jwt: '',
+  // jwt: '',
   rsaPassPhrase: safestr(process.env.rsaPassPhrase),
   rsaPrivateKey: safestr(process.env.rsaPrivateKey),
   rsaPublicKey: safestr(process.env.rsaPublicKey),
@@ -225,23 +222,23 @@ export const TEST_Settings: {
   userIdGoodEmail: 'test@test.com',
 }
 
-export function GenerateSignedJwtToken(
-  email: string = TEST_Settings.userIdGoodEmail,
-  overrides?: Partial<JwtPayload>
-) {
-  return JwtTokenWithEmail(
-    email,
-    TEST_Settings.rsaPrivateKey,
-    TEST_Settings.rsaPassPhrase,
-    overrides
-  )
-}
+// export function GenerateSignedJwtToken(
+//   email: string = TEST_Settings.userIdGoodEmail,
+//   overrides?: Partial<JwtPayload>
+// ) {
+//   return JwtTokenWithEmail(
+//     email,
+//     TEST_Settings.rsaPrivateKey,
+//     TEST_Settings.rsaPassPhrase,
+//     overrides
+//   )
+// }
 
 beforeAll(() => {
   // Process.env.NODE_ENV = 'test'
-  const jwtToken = GenerateSignedJwtToken(TEST_Settings.userIdGoodEmail)
+  // const jwtToken = GenerateSignedJwtToken(TEST_Settings.userIdGoodEmail)
 
-  TEST_Settings.jwt = jwtToken
+  // TEST_Settings.jwt = jwtToken
 
   // MockServer.listen({
   //   // This tells MSW to throw an error whenever it
