+++ src/primitives/map-helper.mts
@@ -1,21 +1,31 @@
-
-export function dictionaryUpsert<K, V>(dict: Map<K, V>, key: K, value: V): void {
+export function dictionaryUpsert<K, V>(
+  dict: Map<K, V>,
+  key: K,
+  value: V
+): void {
   dict.set(key, value)
 }
 
-export function dictionaryUpsertAll<K, V>(dictTarget: Map<K, V>, dictSource: NodeJS.Dict<V>): void {
+export function dictionaryUpsertAll<K, V>(
+  dictTarget: Map<K, V>,
+  dictSource: NodeJS.Dict<V>
+): void {
   for (const [key, value] of Object.entries(dictSource)) {
     dictionaryUpsert(dictTarget, key as K, value)
   }
 }
 
-export function dictionaryFromObject<K extends string | number | symbol, V>(obj: NodeJS.Dict<V>): Map<K, V> {
+export function dictionaryFromObject<K extends string | number | symbol, V>(
+  obj: NodeJS.Dict<V>
+): Map<K, V> {
   const map = new Map<K, V>()
   dictionaryUpsertAll(map, obj)
   return map
 }
 
-export function dictionaryToObject<K extends string | number | symbol, V>(map: Map<K, V>): NodeJS.Dict<V> {
+export function dictionaryToObject<K extends string | number | symbol, V>(
+  map: Map<K, V>
+): NodeJS.Dict<V> {
   const obj: NodeJS.Dict<V> = {}
   for (const [key, value] of map.entries()) {
     obj[key as string] = value
@@ -25,6 +35,6 @@ export function dictionaryToObject<K extends string | number | symbol, V>(map: M
 
 export function processEnvUpsert(dict: NodeJS.Dict<string>): void {
   for (const [key, value] of Object.entries(dict)) {
-    process.env[key] = value
+    // process.env[key] = value
   }
 }
